js/ETH: abstract out ethers
