+++ js/stdlib/ts/module_typecheck.ts
@@ -1,10 +1,13 @@
-// This file only exists to check that modules satisfy an interface
+// This file only exists to typecheck that modules satisfy an interface
 
 import * as ETH from './ETH';
 import * as ETH_compiled from './ETH_compiled';
 import * as CFX from './CFX';
+import ethers from 'ethers';
+import * as cfxers from './cfxers';
+import * as ETH_impl from './ETH_impl';
 
-import { EthLike, EthLikeCompiled } from './ETH_like_interfaces';
+import { EthersLike, EthLike, EthLikeArgs, EthLikeCompiled } from './ETH_like_interfaces';
 
 const _ETH_compiled: EthLikeCompiled = ETH_compiled;
 void(_ETH_compiled);
@@ -14,3 +17,12 @@ void(_ETH);
 
 const _CFX: EthLike = CFX;
 void(_CFX);
+
+const _ethers: EthersLike = ethers;
+void(_ethers);
+
+const _cfxers: EthersLike = cfxers;
+void(_cfxers);
+
+const _ethImpl: EthLikeArgs = ETH_impl;
+void(_ethImpl);
